Moved TrakingModel to trckers module
